refactor: move copyright notice into its own component so it can be shown separate from ts and cs
